+++ scripts/winesapos-install.sh
@@ -1190,6 +1190,9 @@ if [[ "${WINESAPOS_BUILD_CHROOT_ONLY}" == "false" ]]; then
         # Configure the Intel Xe driver to work for the first generation of devices.
         sed -i 's/GRUB_CMDLINE_LINUX="/GRUB_CMDLINE_LINUX="i915.force_probe=!9a49 xe.force_probe=9149 /g' "${WINESAPOS_INSTALL_DIR}"/etc/default/grub
 
+        # Configure USB to not autosuspend.
+        sed -i 's/GRUB_CMDLINE_LINUX="/GRUB_CMDLINE_LINUX="usbcore.autosuspend=-1 /g' "${WINESAPOS_INSTALL_DIR}"/etc/default/grub
+
         efi_partition=2
         if [[ "${WINESAPOS_ENABLE_PORTABLE_STORAGE}" == "true" ]]; then
             efi_partition=3