+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Change Linux LTS to be based on Linux 6.18.
 - Change 'gamescope' package to be the 'gamescope-ogc' package instead.
 - Remove ReiserFS since it is no longer available in the latest Linux LTS.
+- Change USB driver to no longer autosuspend.
 
 ## 4.5.0 - The Ventoy Resize Update